🚚 페이지 단순화를 위해 각 페이지 요소들 컴포넌트화 & 분리
