+++ public/index.html
@@ -5,7 +5,7 @@
     <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
-    <meta name="og:title" content="File-Anywhere" />
+    <meta name="og:title" content="File Anywhere" />
     <meta
       name="og:description"
       content="You can easily send or upload files anywhere and whenever you want."
@@ -14,7 +14,7 @@
       name="description"
       content="You can easily send or upload files anywhere and whenever you want."
     />
-    <title>File-Anywhere</title>
+    <title>File Anywhere</title>
   </head>
   <body>
     <div id="root"></div>
