+++ src/components/Detail/DetailFileList.tsx
@@ -1,0 +1,63 @@
+import React from "react";
+import styled from "styled-components";
+import colors from "styles/colors";
+
+export const DetailFileList = () => {
+  return (
+    <FileList>
+      <FileListItem>
+        <FileItemInfo>
+          <span />
+          <span>logo.png</span>
+        </FileItemInfo>
+        <FileItemSize>10.86KB</FileItemSize>
+      </FileListItem>
+    </FileList>
+  );
+};
+
+const FileList = styled.ul`
+  border-top: 1px solid;
+  border-color: ${colors.grey200};
+  padding: 0;
+  margin: 0;
+  padding: 0 36px;
+  color: ${colors.grey700};
+
+  @media (max-width: 768px) {
+    padding: 0 24px;
+  }
+
+  & > li + li {
+    border-top: 1px solid;
+    border-color: ${colors.grey200};
+  }
+`;
+
+const FileListItem = styled.li`
+  height: 72px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+`;
+
+const FileItemInfo = styled.div`
+  flex-grow: 0;
+  max-width: 50%;
+  flex-basis: 50%;
+  display: flex;
+  align-items: center;
+
+  span:first-child {
+    width: 40px;
+    height: 40px;
+    margin-right: 12px;
+    display: inline-block;
+    background-image: url(/svgs/default.svg);
+    background-size: contain;
+    background-repeat: no-repeat;
+    background-position: center center;
+  }
+`;
+
+const FileItemSize = styled.div``;
